Apply left-arrow-icon to the blog post site title
Lays the icon to the left of the blog post site title link.  One thing I
am not a fan of us when you do go back to the blog index page, the
moving back does not feel as seamless because the site title, while
placed in the same space as the blog index site title link, now jumps to
the left because of the left arrow icon addition.