Update blog dates to include datePublished + dateModified
[issue 73] This has been something I have been wanting to add for a
while.  It's always frustrating when you don't know the date a post was
originally written or updated on.  This is especially true for technical
articles because the date can impact the relevance.  Having said this,
because we are writing CLJS and CLJ posts, the relevance is usually
extended because of the care and consideration and lack of churn within
the ecosystem