Add CustomHeader prop to Layout component
[ISSUE #126] The CustomHeader prop has been added so we can remove the
Header from inside the `main` HTML element and also so we can remove it
from the grid layout and have it follow it's own layout journey.  We did
this to enable us to cubify the header component.

Aside from the above, we also couple adding the Header as a CustomHeader
for the index screen