Clarify that Reagent components default to class components
[ISSUE #150] As of Reagent 1.0.0 Reagent developers can choose if they
want to render a function component or a class component.  This is
important knowledge in the context of this blog post so we update the
intro to stress that the "default" is a class component and even update
the benefits of this post.

The reason we actually remove one of the benefits is because "why can't
reagent render a function component" is no longer a question (because it
can) and likely always could have.  Thus, that point would probably
confuse more than enlighten