Tighten up section - "a pseudoclassical pattern"
[ISSUE #150] With this section I removed some of the overly light
hearted tone I originally had and just made the paras read a little
tigher.  The biggest change is my review of why digging into this is
important, before I made it seem more like an "interesting" thing, but
the truth is that this knowledge (patterns, how to use them and just
being discerning about everything) is an important part of moving out of
the beginner/intermediate phase of programming and into a senior space.