Add filters - dateFilter and w3dateFilter
Added a few commonly used filters.  Also setting an example of how we
can specify the filters file.  Note that there is just 1 index.js file
and we don't have 1 file per function.  That, 1 file per function, is
common in the JS world, but it's an odd convention when there is a low
number of functions in general.  One file will do nicely