Update README - improve quickstart
Specifically callout creating a new dir.  I feel this will help new
developers as it's more inline with what you would actually want to do.
The downside is the top level directory is unknown to us so depending on
what the user does future instructions could be more difficult to grok