Update README - post date formatting
Fought with this for about an hour.  Not in the mood.  The dates comes
from eleventy in two different formats:

- 2019-05-20
- Mon, 14 Jun 2021 01:00:43 GMT

Avoiding the off by one issue is doable in dateFilter, I just need to
think through a way to make it possible without the user having the
think about anything.  Thus, out of lazynesss, just provide the string
formatted as noted in the README and all should be good