update Blog Post - 005 - spelling and clarity
I have not re-run the steps themselves.  The reason is because this post
relies on a few moving pieces and is not something I would normally
write again because it's difficult to maintain, has a lot of
dependencies and doesn't teach fundamentals (which is one of the reasons
it's difficult to maintain).